Use no_std by default
The `std` feature is only needed to enable `impl Write for SmallVec`,
which most users of this crate do not need.  This patch replaces the
enabled-by-default `std` feature with a disabled-by-default `write`
feature.  This decreases the chance that users of this crate will
accidentally depend on libstd when they don't need it.