Implement abomonation support
The `abomonation` crate allows transmutation of types into buffers of
memory and back again.  This is useful for caches, where it makes it
possible to serve read-only references of the item directly out of the
cache memory.

The `abomonation` crate implements support for normal vectors.  This
change adds a new feature (`abomonation`) to the smallvec crate which
provides abomonation support for `SmallVec`, too.